+++ src/imio/smartweb/core/browser/layout.py
@@ -1,10 +1,13 @@
 # -*- coding: utf-8 -*-
 from collective.anysurfer.layout import LayoutPolicy as AnysurferLayoutPolicy
+from eea.facetednavigation.config import ANNO_FACETED_LAYOUT
+from eea.facetednavigation.subtypes.interfaces import IFacetedNavigable
 from imio.smartweb.core.behaviors.minisite import IImioSmartwebMinisite
 from imio.smartweb.core.viewlets.banner import BannerViewlet
 from imio.smartweb.core.viewlets.subsite import BaseSubsiteViewlet
 from plone import api
 from plone.app.layout.globals import layout as base
+from zope.annotation.interfaces import IAnnotations
 
 
 class LayoutPolicy(AnysurferLayoutPolicy):
@@ -18,6 +21,7 @@ class LayoutPolicy(AnysurferLayoutPolicy):
          1. banner image
          2. minisite
          3. subsite
+         4. faceted layout
         """
         context = self.context
         request = self.request
@@ -40,4 +44,10 @@ class LayoutPolicy(AnysurferLayoutPolicy):
         if subsite_viewlet.available():
             body_class += " is-in-subsite"
 
+        # 4. Add faceted class to body if a faceted layout is define
+        if IFacetedNavigable.providedBy(context):
+            faceted_layout = IAnnotations(self.context).get(
+                ANNO_FACETED_LAYOUT, "faceted-block-view"
+            )
+            body_class += f" {faceted_layout}"
         return body_class